feat: move trade execution state variable into trade quote slice fo rmulti-hop progression
