+++ src/state/slices/swappersSlice/selectors.ts
@@ -83,11 +83,6 @@ export const selectSellAmountCryptoPrecision = createSelector(
   swappers => swappers.sellAmountCryptoPrecision,
 )
 
-export const selectTradeExecutionStatus = createSelector(
-  selectSwappers,
-  swappers => swappers.tradeExecutionStatus,
-)
-
 export const selectBuyAssetUsdRate = createSelector(
   selectBuyAsset,
   selectCryptoMarketData,
