+++ src/test/mocks/store.ts
@@ -2,7 +2,7 @@ import { DEFAULT_HISTORY_TIMEFRAME } from 'constants/Config'
 import type { ReduxState } from 'state/reducer'
 import { defaultAsset } from 'state/slices/assetsSlice/assetsSlice'
 import { CurrencyFormats } from 'state/slices/preferencesSlice/preferencesSlice'
-import { MultiHopExecutionStatus } from 'state/slices/swappersSlice/types'
+import { MultiHopExecutionStatus } from 'state/slices/tradeQuoteSlice/types'
 
 const mockApiFactory = <T extends unknown>(reducerPath: T) => ({
   queries: {},
@@ -175,7 +175,6 @@ export const mockStore: ReduxState = {
     sellAssetAccountId: undefined,
     buyAssetAccountId: undefined,
     sellAmountCryptoPrecision: '0',
-    tradeExecutionStatus: MultiHopExecutionStatus.Unknown,
     willDonate: true,
     manualReceiveAddress: undefined,
     manualReceiveAddressIsValidating: false,
@@ -185,5 +184,6 @@ export const mockStore: ReduxState = {
     activeQuoteIndex: undefined,
     confirmedQuote: undefined,
     activeStep: undefined,
+    tradeExecutionStatus: MultiHopExecutionStatus.Unknown,
   },
 }
